backend/requirements.txt
requests==2.26.0
Django==3.2.3
PyJWT==2.1.0
djangorestframework==3.12.4
django-cors-headers==3.13.0
djoser==2.1.0
djangorestframework-simplejwt==4.7.2
django-filter==23.1
gunicorn==20.1.0
webcolors==1.11.1
psycopg2-binary==2.9.3
Pillow==9.0.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dotenv==1.0.1
PyYAML==6.0